Add 2 sample projects for GoogleMobileVision pod.
BarcodeDetector: This demo demonstrates basic barcode detection and
integration with AVFoundation. The app highlights barcodes in the
camerapreview frame.

MultiDetector: This demo demonstrates how to use the MultiDataOutput in
GoogleMVDataOutput pod to integrate both face and barcode detectors with
the video pipeline. The app draws overlay on top of detected barcodes
and faces.

Change-Id: I6d95a39a6294eb3a29baedb4d98d29aa864da29c